refactor: reorganize types and composables structure
Consolidates shared type definitions into centralized locations to improve code organization and reusability across layers.

Moves type definitions from layer-specific and app-specific directories to shared directories, enabling better access patterns and reducing duplication.

Relocates globe-related composables from component directory to app composables for proper auto-import support and better separation of concerns.

Updates import paths throughout to reflect new structure using path aliases for cleaner module resolution.
