+++ app/types/index.ts
@@ -1,3 +1,3 @@
-export * from '#layers/dashboard/app/types'
-export * from './link'
-export * from './traffic'
+export * from '#layers/dashboard/shared/types'
+export * from '#shared/types/link'
+export * from '#shared/types/traffic'
