+++ layers/dashboard/app/composables/useGlobeData.ts
@@ -1,7 +1,10 @@
 import type { GlobeConfig } from '#layers/dashboard/app/composables/useD3Globe'
 import type { ComputedRef, Ref, ShallowRef } from 'vue'
 import type { AreaData, ColoData, CurrentLocation, GeoJSONData, LocationData, TrafficEventParams } from '@/types'
+import { computed, ref, shallowRef } from 'vue'
 import { watchDeep } from '@vueuse/core'
+import { useDashboardRealtimeStore } from './realtime'
+import { useAPI } from '@/utils/api'
 
 interface RawLocationData {
   latitude: number
