+++ shared/types/link.ts
