+++ shared/types/traffic.ts
